ShoesOnlineStore/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.2.7
cron-descriptor==1.4.0
Django==4.2.3
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-ckeditor==6.6.1
django-cors-headers==4.2.0
django-countries==7.5.1
django-filter==23.2
django-js-asset==2.1.0
django-redis==5.3.0
django-timezone-field==5.1
djangorestframework==3.14.0
drf-spectacular==0.26.4
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
isodate==0.6.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kavenegar==1.1.2
kombu==5.3.1
lxml==4.9.3
model-bakery==1.12.0
packaging==23.1
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
referencing==0.30.2
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rpds-py==0.9.2
six==1.16.0
sqlparse==0.4.4
supervisor==4.2.5
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
